tests/wpt/tests/html/browsers/the-windowproxy-exotic-object/windowproxy-prevent-extensions.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>[[PreventExtensions]] on a WindowProxy object should return false</title>
<link rel="help" href="https://html.spec.whatwg.org/multipage/window-object.html#windowproxy-preventextensions">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
"use strict";

test(() => {
  assert_throws_js(TypeError, () => {
    Object.preventExtensions(window);
  });
}, "Object.preventExtensions throws a TypeError");

test(() => {
  assert_false(Reflect.preventExtensions(window));
}, "Reflect.preventExtensions returns false");
</script>
